Split RViz headers into a separate directory.
